+++ src/wpc/s11.c
@@ -415,9 +415,9 @@ static WRITE_HANDLER(pia2b_w) {
            locals.pseg[locals.digSel].w = core_bcd2seg7[data&0x0f];
       locals.segments[20+locals.digSel].w |=
            locals.pseg[20+locals.digSel].w = core_bcd2seg7[data>>4];
-      core_write_pwm_output_8b(CORE_MODOUT_SEG0 + locals.digSel * 16, core_bcd2seg7[data & 0x0f]);
+      core_write_pwm_output_8b(CORE_MODOUT_SEG0 + locals.digSel * 16, (UINT8)core_bcd2seg7[data & 0x0f]);
       core_write_pwm_output_8b(CORE_MODOUT_SEG0 + locals.digSel * 16 + 8, core_bcd2seg7[data & 0x0f] >> 8);
-      core_write_pwm_output_8b(CORE_MODOUT_SEG0 + (20 + locals.digSel) * 16, core_bcd2seg7[data >> 4]);
+      core_write_pwm_output_8b(CORE_MODOUT_SEG0 + (20 + locals.digSel) * 16, (UINT8)core_bcd2seg7[data >> 4]);
       core_write_pwm_output_8b(CORE_MODOUT_SEG0 + (20 + locals.digSel) * 16 + 8, core_bcd2seg7[data >> 4] >> 8);
     }
     else
